# REST_APIs_Design/openapi.yaml
openapi: 3.0.1
info:
   title: Film Manager
   description: REST APIs for the interaction with the Film Manager service.
   termsOfService: 'http://swagger.io/terms/'
   license:
      name: Apache 2.0
      url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
   version: 1.0.0
...
servers:
   - url: 'https://localhost/v1'
   - url: 'http://localhost/v1'
...
paths:
...
   '/api/films/public':
      get:
         tags:
            - films/public
         summary: Retrieve the public films
         description: >-
            The public films (i.e., the films that are visible for all the users of the service) are retrieved. This operation does not require authentication. A pagination mechanism is implemented to limit the size of messages.
         operationId: getPublicFilms
         parameters:
            - in: query
              name: pageNo
              required: false
              schema:
                 type: integer
              description: 'The id of the requested page (if absent, the first page is returned)'
         responses:
            '200':
               description: Successful operation
               content:
                  application/json:
                     schema:
                        type: object
                        properties:
                           totalPages:
                              type: integer
                           currentPage:
                              type: integer
                           totalItems:
                              type: integer
                           films:
                              type: array
                              items:
                                 $ref: '#/components/schemas/Film'
                           next:
                              type: string
                              format: uri
            '500':
               description: Internal Server Error
               content: {}
...
   '/api/films/public/{filmId}/reviews/{reviewId}/drafts':
      get:
         tags:
            - 'films/public/{filmId}/reviews/{reviewId}/drafts'
         security:
            - cookieAuth: []
         summary: Retrieve all drafts that are opened/closed for a review with ID reviewId
         description: >-
            All the drafts of the review with ID reviewId are retrieved.
            This operation require authentication.
         operationId: getDrafts
         parameters:
            - name: filmId
              in: path
              description: ID of the film whose reviews must be retrieved
              required: true
              schema:
                 type: integer
                 format: int64
            - name: reviewId
              in: path
              description: ID of the review which must be retrieved
              required: true
              schema:
                 type: integer
                 format: int64
            - in: query
              name: pageNo
              required: false
              schema:
                 type: integer
              description: 'The id of the requested page (if absent, the first page is returned)'
         responses:
            '200':
               description: Successful operation
               content:
                  application/json:
                     schema:
                        type: array
                        items:
                           type: object
                           properties:
                              totalPages:
                                 type: integer
                              currentPage:
                                 type: integer
                              totalItems:
                                 type: integer
                              reviews:
                                 type: array
                                 items:
                                    $ref: '#/components/schemas/Draft'
                              next:
                                 type: string
                                 format: uri
            '401':
               description: Unauthorized
               content: {}
            '404':
               description: Not Found
               content: {}
            '500':
               description: Internal Server Error
               content: {}
      post:
         tags:
            - 'films/public/{filmId}/reviews/{reviewId}/drafts'
         security:
            - cookieAuth: []
         summary: Create a new draft for a review with ID reviewId
         description:
            Create a drafts for review with ID reviewId. This operation only allows opening a draft if no other
            drafts are opened. This operation can be performed only by the reviewer participating in
            the coop review.
         operationId: createDraft
         parameters:
            - name: filmId
              in: path
              description: ID of the film whose review must be completed
              required: true
              schema:
                 type: integer
                 format: int64
            - name: reviewId
              in: path
              description: ID of the review which must be retrieved
              required: true
              schema:
                 type: integer
                 format: int64
         requestBody:
            description: The properties of the Review object that must be updated
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/Draft'
         responses:
            '204':
               description: No Content
               content: {}
            '400':
               description: Bad Request
               content: {}
            '401':
               description: Unauthorized
               content: {}
            '403':
               description: Forbidden
               content: {}
            '404':
               description: Not Found
               content: {}
            '418':
               description: I'm a Teapot
               content: {}
            '500':
               description: Internal Server Error
               content: {}
   '/api/films/public/{filmId}/reviews/{reviewId}/drafts/{draftId}':
      get:
         tags:
            - 'films/public/{filmId}/reviews/{reviewId}/drafts/{draftId}'
         security:
            - cookieAuth: []
         summary: Retrieve the draft with ID draftId for the review with ID reviewId
         description: >-
            The drafts with ID draftId of the review with ID reviewId is retrieved.
            This operation require authentication.
         operationId: getDraft
         parameters:
            - name: filmId
              in: path
              description: ID of the film whose reviews must be retrieved
              required: true
              schema:
                 type: integer
                 format: int64
            - name: reviewId
              in: path
              description: ID of the review which must be retrieved
              required: true
              schema:
                 type: integer
                 format: int64
            - name: draftId
              in: path
              description: ID of the draft which must be retrieved
              required: true
              schema:
                 type: integer
                 format: int64
         responses:
            '200':
               description: Successful operation
               content:
                  application/json:
                     schema:
                        $ref: '#/components/schemas/Draft'
            '401':
               description: Unauthorized
               content: {}
            '404':
               description: Not Found
               content: {}
            '500':
               description: Internal Server Error
               content: {}
      put:
         tags:
            - 'films/public/{filmId}/reviews/{reviewId}/drafts/{draftId}'
         security:
            - cookieAuth: []
         summary: Judge the draft with ID draftId for the review with ID reviewId
         description:
            Judge the draft with ID draftId for the review with ID reviewId. This operation can be
            performed only by the reviewer participating in the coop review.
         operationId: judgeDraft
         parameters:
            - name: filmId
              in: path
              description: ID of the film whose review must be completed
              required: true
              schema:
                 type: integer
                 format: int64
            - name: reviewId
              in: path
              description: ID of the review which must be retrieved
              required: true
              schema:
                 type: integer
                 format: int64
            - name: draftId
              in: path
              description: ID of the draft which must be retrieved
              required: true
              schema:
                 type: integer
                 format: int64
         requestBody:
            description: The properties of the Review object that must be updated
            content:
               application/json:
                  schema:
                     $ref: '#/components/schemas/Judgment'
         responses:
            '204':
               description: No Content
               content: {}
            '400':
               description: Bad Request
               content: {}
            '401':
               description: Unauthorized
               content: {}
            '403':
               description: Forbidden
               content: {}
            '404':
               description: Not Found
               content: {}
            '418':
               description: I'm a Teapot
               content: {}
            '500':
               description: Internal Server Error
               content: {}
...
components:
   schemas:
...
      Film:
         title: Film
         description: JSON Schema for Film data structures
         type: object
         additionalProperties: false
         properties:
           "$schema":
             type: string
             description: path to the JSON Schema to use for the validation of the data structure
           id:
             type: integer
             description: unique identifier for the film
           title:
             type: string
             description: textual title of the film
           owner:
             type: integer
             description: unique identifier of the film owner
           private:
             type: boolean
             description: true if only the film owner can have access to it, false otherwise
             default: true
           watchDate:
             type: string
             format: date
             description: the date when the film was watched by the owner
           rating:
             type: integer
             description: the rating the owner has given to the film
             minimum: 1
             maximum: 10
           favorite:
             type: boolean
             description: 'true if the film is among the favourite ones of the owner, false
               otherwise '
             default: false
         required:
         - id
         - title
         - owner
         - private
...
      Draft:
         title: draft
         description: JSON Schema for Draft data structures
         type: object
         additionalProperties: false
         properties:
           "$schema":
             type: string
             description: path to the JSON Schema to use for the validation of the data structure
           id:
             type: integer
             description: unique identifier for the draft review
           reviewId:
             type: integer
             description: unique identifier of the the review for which this draft has been
               issued
           reviewerId:
             type: integer
             description: unique identifier of the the user who has created this draft review
           open:
             type: boolean
             description: true if the draft review is open, false otherwise
             default: true
           proposedRating:
             type: integer
             description: the proposed rating the reviewer has given to the film
             minimum: 1
             maximum: 10
           proposedReview:
             type: string
             description: a textual description of the proposed review
             maxLength: 1000
           judgment:
            type: array
            items:
               $ref: '#/components/schemas/Judgment'
         required:
         - reviewId
         - reviewerId
         - proposedRating
         - proposedReview
      Judgment:
         title: Judgment
         description: JSON Schema for Judgment data structures
         type: object
         additionalProperties: false
         properties:
           "$schema":
             type: string
             description: path to the JSON Schema to use for the validation of the data structure
           draftId:
             type: integer
             description: unique identifier of the the draft review for which this judgment
               has been issued
           reviewerId:
             type: integer
             description: unique identifier of the the user who has created this judgment
           agree:
             type: boolean
             description: true if the judgment is positive, false otherwise
             default: true
           comment:
             type: string
             description: a textual description of the disagreement reason
             maxLength: 1000
         required:
         - draftId
         - reviewerId
   securitySchemes:
      cookieAuth:
         description: JWT to identify the authenticated user.
         type: apiKey
         in: cookie
         name: connect.sid